djangoecommerce/catalog/requirements.txt
django-test-without-migrations==0.4
